Bloom: Reduce dynamic range during the first downsample pass
Changed the sampler to have a black border color, as this helps to avoid
bloom getting more intensive closer to the viewport edges.

Previously the texture samples were clamped to the border, and the
shader was reading the same high intensity values.
This lead to having flickering under the camera motion (the bloom was
suddenly getting more bright before completely disappearing).